requirements.txt
Django==1.11.5
django-rest-framework==0.1.0
djangorestframework==3.6.4
